requirements.txt
streamlit
openai
fal-client
requests
python-dotenv
markdown
markdown-it-py
Pillow
playwright
openpyxl
